Potentially fixes a problem with accessibility.
When performing a search, the status of the results only shows as an icon, which is unreadable with screen readers.
This fix attempts to correct this by hiding the icon to screen readers, and adding the status text so it is readable.
An alternative method would be to use aria-label, but I chose to insert the name as part of the text for the rare occasion that ARIA isn't available or supported.